+++ programs/bpf_loader/src/lib.rs
@@ -16,7 +16,7 @@ use log::{log_enabled, trace, Level::Trace};
 use solana_measure::measure::Measure;
 use solana_program_runtime::{
     ic_logger_msg, ic_msg,
-    instruction_processor::{Executor, InstructionProcessor},
+    instruction_processor::Executor,
     invoke_context::{ComputeMeter, InvokeContext},
     log_collector::LogCollector,
     stable_log,
@@ -485,7 +485,7 @@ fn process_loader_upgradeable_instruction(
                 .iter()
                 .map(|seeds| Pubkey::create_program_address(*seeds, caller_program_id))
                 .collect::<Result<Vec<Pubkey>, solana_sdk::pubkey::PubkeyError>>()?;
-            InstructionProcessor::native_invoke(invoke_context, instruction, signers.as_slice())?;
+            invoke_context.native_invoke(instruction, signers.as_slice())?;
 
             // Load and verify the program bits
             let executor = create_executor(
